feature: sending JWT token
