+++ travis/controllers/mainController.js
@@ -1,6 +1,4 @@
 const UserModel = require('../models/userModel');
-// const passport = require('passport');
-// const LocalStrategy = require('passport-local').Strategy;
 
 exports.getMainPage = (req, res) => {
   try {
